# room-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Layanan Ruangan",
        "description": "Dokumentasi API interaktif untuk layanan ruangan",
        "version": "1.0.0"
    },
    "paths": {
        "/api/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Ambil daftar semua ruangan",
                "operationId": "f36201d4142b0b6e142bcd5e5d0bcc1e",
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Room"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Tambah ruangan baru",
                "operationId": "5926474104d331e075299a5b9d2f7416",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "capacity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ruang Kelas A"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "facilities": {
                                        "type": "string",
                                        "example": "AC, LCD Projector"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ruangan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Lihat detail ruangan",
                "operationId": "2fb25667a180bad0871584c60fa392fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ruangan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ruangan tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update data ruangan",
                "operationId": "a16e22295fb7ccec81d4a685ab6edf52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ruangan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ruang Rapat Besar"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "facilities": {
                                        "type": "string",
                                        "example": "AC, Sound System, Projector"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ruangan tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Hapus ruangan",
                "operationId": "9f216809b68e2ce14c857a26768e9228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ruangan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ruangan berhasil dihapus"
                    },
                    "404": {
                        "description": "Ruangan tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Room": {
                "required": [
                    "name",
                    "capacity"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ruang Rapat 1"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 20
                    },
                    "facilities": {
                        "type": "string",
                        "example": "Proyektor, AC, Whiteboard"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Rooms",
            "description": "Rooms"
        }
    ]
}